requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
requests>=2.28.0
pycryptodome>=3.15.0
pywin32>=305
